Redistribute C++ coverage tests into feature-aligned test files (#782)
Move tests from test_cpp_coverage_tier1.py and test_cpp_coverage_torch.py
into existing files by topic: repr tests to test_tostring, texture shape to
test_textures, Shape/SignatureBuilder to test_shape, buffer view indexing to
test_buffer_views, torch dtypes to test_torch_dtypes, raw dispatch to
test_raw_dispatch, DiffPair to test_torch_marshall_gaps, and benchmarks
to a new standalone test_torch_benchmarks.py.

Made-with: Cursor
